third_party/blink/web_tests/platform/generic/fast/dom/HTMLTemplateElement/cloneNode-expected.txt

Test that template contents are not cloned when the template element is cloned

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS template.content.childNodes.length is 1
PASS clone.content.childNodes.length is 1
PASS clone.outerHTML is template.outerHTML
PASS clone.content.firstChild is not template.content.firstChild
PASS clone.content is not template.content
PASS clone.firstChild.tagName is "DIV"
PASS clone.firstChild is not div
PASS successfullyParsed is true

TEST COMPLETE
